switched to user_playlist_tracks attribute, adding every track in a playlist
In the original code, the script stops short once it hits the 100th song. In this code, it grabs every song from a playlist before it returns the listed dictionaries.

This, therefore, ensures you will be able to grab every single image from the playlist (in the original script, this isn't possible if you had a large playlist).

Note: get_artist is still limited.